Split path impls into their own files
